src/main/java/com/thoughtworks/tdd/parklinglot/shell/controller/ParkingController.java
package com.thoughtworks.tdd.parklinglot.shell.controller;

import com.thoughtworks.tdd.parklinglot.core.Car;
import com.thoughtworks.tdd.parklinglot.core.ParkingBoy;
import com.thoughtworks.tdd.parklinglot.core.Receipt;
import com.thoughtworks.tdd.parklinglot.shell.io.Request;
import com.thoughtworks.tdd.parklinglot.shell.io.Response;

import java.util.UUID;

public class ParkingController implements BaseController {
    private final Request request;
    private final Response response;
    private ParkingBoy parkingBoy;

    public ParkingController(Request request, Response response, ParkingBoy parkingBoy) {

        this.request = request;
        this.response = response;
        this.parkingBoy = parkingBoy;
    }


    @Override
    public String process() {
        String cardNum = request.getCommand();

        Receipt receipt = parkingBoy.park(new Car(cardNum));

        response.send("停车成功，您的小票是：\n" +
                receipt.toString());
        return "forward:main";
    }
}
